add a separate file for univariate distribution
